Move bin to be compilable
