+++ src/bin/lint.js
