+++ src/bin/toAsciidoc.js
@@ -2,11 +2,11 @@
 /* eslint-disable no-console */
 
 const { transform } = require('./helper');
-const { State } = require('../src/');
-const html = require('../src/html');
+const { State } = require('../');
+const asciidoc = require('../asciidoc');
 
 transform(document => {
-    const state = State.create(html);
+    const state = State.create(asciidoc);
     const output = state.serializeDocument(document);
 
     console.log(output);
