+++ src/bin/toHTML.js
@@ -2,11 +2,11 @@
 /* eslint-disable no-console */
 
 const { transform } = require('./helper');
-const { State } = require('../src/');
-const markdown = require('../src/markdown');
+const { State } = require('../');
+const html = require('../html');
 
 transform(document => {
-    const state = State.create(markdown);
+    const state = State.create(html);
     const output = state.serializeDocument(document);
 
     console.log(output);
